+++ crates/jsonschema-py/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Resolving external references that nested inside local references. [#671](https://github.com/Stranger6667/jsonschema/issues/671)
 
+### Performance
+
+- Faster JSON pointer resolution.
+
 ## [0.28.1] - 2024-12-31
 
 ### Fixed